requirements.txt
alembic==1.9.2
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
autoflake==1.7.8
black==22.12.0
certifi==2022.12.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
Flake8-pyproject==1.2.2
flake8-return==1.2.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
identify==2.5.13
idna==3.4
iniconfig==2.0.0
isort==5.11.4
Mako==1.2.4
MarkupSafe==2.1.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==23.0
pathspec==0.10.3
pep8-naming==0.13.3
platformdirs==2.6.2
pluggy==1.0.0
pre-commit==2.21.0
psycopg2-binary==2.9.5
pyaes==1.6.1
pycodestyle==2.9.1
pydantic==1.10.4
pyflakes==2.5.0
Pyrogram==2.0.97
PySocks==1.7.1
pytest==7.2.1
pytest-asyncio==0.20.3
python-dotenv==0.21.0
python-telegram-bot==20.0
PyYAML==6.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
TgCrypto==1.2.5
tomli==2.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
virtualenv==20.17.1
